api_design/FastAPI/clean_architecture/level-2/requirements.txt
alembic==1.16.0
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.27.0
click==8.2.1
fastapi==0.100.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.0.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
